(#4377) Status bar graph resize and interaction updates
Bar graph does not resize and redraw when the view port of
the browser is scaled up or down. Graph does not allow for
click events to route to individual reports. Add custom
window resize function which updates the view port's width.
Add custom click events that link to reports.